Enable Travis-CI for darwin-free
There's not really a test suite (patches welcome).  All this does is
make sure the program builds on OS X, and basically whatever the current
default OS X image is on Travis-CI.  Please feel free to submit PRs to
extend this file and support testing on specific versions or in
Homebrew.